Revert "[OP] add oneDNN relu unit tests - WIP."
This reverts commit fb397d985eb99872f2bc652ca0bc28a5917605d5.
